Add heading tag helper to main helper manifest
